Making the Errors Folder if its not there
